checkpoint: liquid glass demo (native NSGlassEffectView refraction + shader effects) plus in-flight agent work
- src/bin/liquid_glass_demo.rs: Spotlight-style liquid glass POC — real
  macOS 26 NSGlassEffectView/NSGlassEffectContainerView refraction hosted
  below GPUI's Metal view, spring-driven morphs, SDF/marching-squares
  fallback, per-shape shader_effect overlays (Tab cycles), native drag.
- Cargo.toml: liquid-glass-demo bin target.
- Remaining files: in-flight work from concurrent agent sessions,
  committed as-is per checkpoint request (cargo fmt applied to satisfy
  the pre-commit hook).

+++ src/design_contract/mod.rs
@@ -4453,7 +4453,7 @@ pub fn checked_in_design_bundle() -> Result<DesignTokenBundle, String> {
             TokenStage::Resolved,
             Some("--sk-arg-input-caret-height"),
             TokenValue::Length {
-                value: arg_caret_height as f64,
+                value: arg_caret_height,
             },
             None,
             false,
